third_party/blink/web_tests/external/wpt/mathml/relations/html5-tree/tabindex-focus-001.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] invalid tabindex attribute does not make the element focusable
  assert_equals: <math href=#> should be focusable expected Element node <math href="#" data-focusable="true"></math> but got Element node <body><math></math>\n<math href="#" data-focusable="true">...
Harness: the test ran to completion.
